move scribbot stuff to its own "sub-struct"
